% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Con_in.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion Matrix}
\usage{
confusion_matrix(X, y, beta_hat, cut = 0.5)
}
\arguments{
\item{X}{Matrix of predictors}

\item{y}{Binary response variable (0 or 1)}

\item{beta_hat}{Coefficient vector}

\item{cut}{Cut off value}
}
\value{
Confidence Matrix
}
\description{
Confusion matrix with a cut-off value for predictions at 0.5. Predictions above 0.5 are assigned a 1, below 0.5 are assigned 0.
}
