% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Con_in.R
\name{prevalencegrid}
\alias{prevalencegrid}
\title{Prevalence Grid}
\usage{
prevalencegrid(X, y, beta_hat, cut = seq(0.1, 0.9, by = 0.1))
}
\arguments{
\item{X}{Matrix of predictors}

\item{y}{Binary response variable (0 or 1)}

\item{beta_hat}{Coefficient vector}

\item{cut}{Vector of cut-off values}
}
\value{
Matrix of prevalence values for each cut-off
}
\description{
Calculates prevalence for different cut-off values and returns a matrix.
}
